correct compare_blobs type

+++ ngraph/test/util/engine/ie_engines.cpp
@@ -89,7 +89,7 @@ namespace
             return compare_blobs<float>(computed, expected, tolerance_bits);
             break;
         case InferenceEngine::Precision::FP64:
-            return compare_blobs<float>(computed, expected, tolerance_bits);
+            return compare_blobs<double>(computed, expected, tolerance_bits);
             break;
         case InferenceEngine::Precision::I8:
             return compare_blobs<int8_t>(computed, expected, tolerance_bits);
